Moved controllers into compareMe.fx package
