+++ src/compareMe/Main.java
@@ -10,7 +10,7 @@ public class Main extends Application {
 
     @Override
     public void start(Stage primaryStage) throws Exception{
-        Parent root = FXMLLoader.load(getClass().getResource("fxml/CompareMeMain.fxml"));
+        Parent root = FXMLLoader.load(getClass().getResource("fx/CompareMeMain.fxml"));
         primaryStage.setTitle("Compare Me");
         primaryStage.setScene(new Scene(root));
         primaryStage.show();
