+++ src/compareMe/fx/ManualDeleterController.java
@@ -1,4 +1,4 @@
-package compareMe;
+package compareMe.fx;
 
 import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.SimpleBooleanProperty;
